+++ api/internal/helloworld/v1/internal/helloworld/in_memory_repository_test.go
@@ -1,4 +1,4 @@
-package v1
+package helloworld
 
 import (
 	"context"
